# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Padel Score Recording API
  version: 1.0.0
  description: API for managing padel players, matches, and head-to-head statistics.

servers:
  - url: http://localhost:8080

paths:
  /players:
    post:
      summary: Create a new player
      operationId: postPlayer
      x-eov-operation-handler: v1/handlers/entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Player created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
        '400':
          description: Missing player name

    get:
      summary: Get all players
      operationId: getPlayers
      x-eov-operation-handler: v1/handlers/entities
      responses:
        '200':
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string

  /players/{id}:
    get:
      summary: Get a player by ID
      operationId: getPlayer
      x-eov-operation-handler: v1/handlers/entities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
        '404':
          description: Player not found
    delete:
      summary: Delete a player by ID
      operationId: deletePlayer
      x-eov-operation-handler: v1/handlers/entities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Player deleted
        '404':
          description: Player not found

  /matches:
    post:
      summary: Create a new match
      operationId: postMatch
      x-eov-operation-handler: v1/handlers/entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [player1, player2, winner, score]
              properties:
                player1:
                  type: string
                player2:
                  type: string
                winner:
                  type: string
                score:
                  type: string
                date:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Match created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  player1:
                    type: string
                  player2:
                    type: string
                  winner:
                    type: string
                  score:
                    type: string
                  date:
                    type: string
                    format: date-time
        '400':
          description: Missing match details

    get:
      summary: Get all matches (most recent first)
      operationId: getMatches
      x-eov-operation-handler: v1/handlers/entities
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    player1:
                      type: string
                    player2:
                      type: string
                    winner:
                      type: string
                    score:
                      type: string
                    date:
                      type: string
                      format: date-time

  /head2head:
    get:
      summary: Get head-to-head statistics between two players
      operationId: getHeadToHead
      x-eov-operation-handler: v1/handlers/entities
      parameters:
        - name: player1
          in: query
          required: true
          schema:
            type: string
        - name: player2
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Head-to-head stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  player1:
                    type: string
                  player2:
                    type: string
                  totalMatches:
                    type: integer
                  player1Wins:
                    type: integer
                  player2Wins:
                    type: integer